+++ src/api/game/userBlock.ts
@@ -1,0 +1,2 @@
+// /api/relationships/block
+// { user: id }