+++ src/api/game/userUnblock.ts
@@ -1,0 +1,2 @@
+// /api/relationships/remove
+// { user: id }