+++ src/index.ts
@@ -1,4 +1,6 @@
 export { default as Client } from "./model/client"
 export * as api              from "./api/mod"
+export *                     from "./util/mod"
+
 export * as CONSTANTS        from "./constants"
 export type * as Types       from "./types"